requirements.txt
streamlit>=1.33
pandas>=2.0
openpyxl>=3.1
plotly>=5.18
pyyaml>=6.0
